refactor(lsp): extract completion handlers into language/completion.rs
Move completion-related functionality from the monolithic server_impl/mod.rs
into a dedicated completion module for better code organization.

Moved to language/completion.rs:
- handle_completion: Main completion request handler
- handle_completion_cancellable: Completion with cancellation support
- format_type_for_detail: Type formatting helper
- degrade_snippet_to_plaintext: Snippet degradation for unsupported clients
- lexical_complete: Fallback completion when AST unavailable

Removed unused imports from mod.rs:
- CompletionProvider (now used in completion.rs)
- TypeInferenceEngine (now used in completion.rs)
- CompletionItemKind (now used in completion.rs)

All 279 lib tests + 17 completion tests + 14 hover tests + 4 definition
tests passing.
