+++ crates/perl-parser/src/lsp/server_impl/language/mod.rs
@@ -2,8 +2,9 @@
 //!
 //! This module organizes LSP language features into focused submodules:
 //! - hover: Hover information and signature help
-//! - (future) completion: Code completion
+//! - completion: Code completion with cancellation support
 //! - (future) navigation: Go-to-definition, references, etc.
 //! - (future) symbols: Document and workspace symbols
 
+mod completion;
 mod hover;
